perf: move `templateNames` into a separate chunk to decrease parsing time
