+++ packages/nuxi/src/main.ts
@@ -15,7 +15,7 @@ import { runCommand } from './run'
 import { setupGlobalConsole } from './utils/console'
 import { checkEngines } from './utils/engines'
 import { debug, logger } from './utils/logger'
-import { templateNames } from './utils/templates'
+import { templateNames } from './utils/templates/names'
 
 // globalThis.crypto support for Node.js 18
 if (!globalThis.crypto) {
@@ -65,7 +65,7 @@ const _main = defineCommand({
       await backgroundTasks
     }
 
-    if (command === 'add' && ctx.rawArgs[1] && templateNames.includes(ctx.rawArgs[1])) {
+    if (command === 'add' && ctx.rawArgs[1] && templateNames.includes(ctx.rawArgs[1] as typeof templateNames[number])) {
       logger.warn(`${colors.yellow('Deprecated:')} Using ${colors.cyan('nuxt add <template> <name>')} is deprecated.`)
       logger.info(`Please use ${colors.cyan('nuxt add-template <template> <name>')} instead.`)
       const addTemplate = await import('./commands/add-template').then(m => m.default || m)
